deploy/flask/requirements.txt
ccxt==1.95.14
Flask==2.1.3
mysql-connector-python==8.0.18